google_maps_mcp/requirements.txt
fastmcp>=0.1.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
